fix(home): fixes relay type error; refactors to be simpler
